+++ src/mark_atom.cpp
@@ -30,4 +30,8 @@ void register_mark_macro() {
     s_registered = true;
 }
 
+void reset_mark_macro() {
+    s_registered = false;
+}
+
 }  // namespace microtex